requirements/base.txt
# File: FlaskVerseHub/requirements/base.txt

# Flask Core
Flask>=2.3.0,<3.0.0
Werkzeug>=2.3.0,<3.0.0
Jinja2>=3.1.0,<4.0.0
MarkupSafe>=2.1.0,<3.0.0
itsdangerous>=2.1.0,<3.0.0
click>=8.1.0,<9.0.0

# Database
SQLAlchemy>=2.0.0,<2.1.0
Flask-SQLAlchemy>=3.0.0,<4.0.0
Flask-Migrate>=4.0.0,<5.0.0
alembic>=1.12.0,<2.0.0

# Authentication & Security
Flask-Login>=0.6.0,<1.0.0
Flask-JWT-Extended>=4.5.0,<5.0.0
PyJWT>=2.8.0,<3.0.0
bcrypt>=4.0.0,<5.0.0
cryptography>=41.0.0,<42.0.0

# Forms & Validation
Flask-WTF>=1.1.0,<2.0.0
WTForms>=3.0.0,<4.0.0
email-validator>=2.0.0,<3.0.0

# Web Features
Flask-Mail>=0.9.1,<1.0.0
Flask-Caching>=2.0.0,<3.0.0
Flask-SocketIO>=5.3.0,<6.0.0
Flask-CORS>=4.0.0,<5.0.0
python-socketio>=5.8.0,<6.0.0
python-engineio>=4.7.0,<5.0.0

# API Development
marshmallow>=3.20.0,<4.0.0
marshmallow-sqlalchemy>=0.29.0,<1.0.0
graphene>=3.3.0,<4.0.0
graphene-sqlalchemy>=3.0.0,<4.0.0
apispec>=6.3.0,<7.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
redis>=4.6.0,<5.0.0
celery>=5.3.0,<6.0.0

# Date & Time
python-dateutil>=2.8.0,<3.0.0
pytz>=2023.3